chore(cosmwasm): use `set_contract_meta` for migration
